+++ build.washingtonpost.com/components/Markdown/Components/headers.js
@@ -8,6 +8,7 @@ export const Header = styled("h1", {
   fontFamily: "$headline",
   color: "$primary",
   fontWeight: "$bold",
+  marginBlock: 0,
   "@sm": {
     fontSize: "$250",
   },