+++ build.washingtonpost.com/pages/playroom.js
@@ -15,138 +15,6 @@ import LZString from "lz-string";
 import { ErrorBoundary } from "react-error-boundary";
 import Link from "~/components/Typography/link";
 
-const lightTheme = {
-  palette: {
-    activeText: "#166dfc",
-    defaultText: "#666666",
-    inactiveText: "#e9e9e9",
-    activeBackground: "#dde6f2",
-    defaultBackground: "#ffffff",
-    inputBackground: "#ffffff",
-    accent: "#166dfc",
-    errorBackground: "#ffffff",
-    errorForeground: "#f27b81",
-  },
-  syntax: {
-    plain: "#111111",
-    comment: {
-      color: "#999",
-      fontStyle: "italic",
-    },
-    keyword: "#166dfc",
-    tag: "#498a0c",
-    punctuation: "#111111",
-    definition: "#092c65",
-    property: "#166dfc",
-    static: "#b16e00",
-    string: "#498a0c",
-  },
-  typography: {
-    bodyFont:
-      '-apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol"',
-    monoFont:
-      '"Fira Mono", "DejaVu Sans Mono", Menlo, Consolas, "Liberation Mono", Monaco, "Lucida Console", monospace',
-    fontSize: "14px",
-    lineHeight: "1.4",
-  },
-};
-
-const darkTheme = {
-  palette: {
-    activeText: "#1761f1",
-    defaultText: "#868585",
-    inactiveText: "#e9e9e9",
-    activeBackground: "#0f1218",
-    defaultBackground: "#020202",
-    inputBackground: "#020202",
-    accent: "#1761f1",
-    errorBackground: "#020202",
-    errorForeground: "#f27b81",
-  },
-  syntax: {
-    plain: "#f3f3f3",
-    comment: {
-      color: "#f3f3f3",
-      fontStyle: "italic",
-    },
-    keyword: "#166dfc",
-    tag: "#517c2a",
-    punctuation: "#f3f3f3",
-    definition: "#cdd4df",
-    property: "#166dfc",
-    static: "#906629",
-    string: "#517c2a",
-  },
-  typography: {
-    bodyFont:
-      '-apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol"',
-    monoFont:
-      '"Fira Mono", "DejaVu Sans Mono", Menlo, Consolas, "Liberation Mono", Monaco, "Lucida Console", monospace',
-    fontSize: "14px",
-    lineHeight: "1.4",
-  },
-};
-
-const sandboxGlobalcss = Kit.globalCss({
-  ".sp-preview-iframe": {
-    background: Kit.theme.colors.gray500,
-    border: 0,
-    width: "100%",
-  },
-  ".sp-layout": {
-    minHeight: 300,
-    borderColor: "transparent",
-    borderBottomLeftRadius: "0 !important",
-    borderBottomRightRadius: "0 !important",
-    background: "$secondary !important",
-  },
-
-  ".sp-code-editor": {
-    py: "var(--sp-space-2)",
-    background: "$secondary",
-    width: "50vw",
-    height: "100vh",
-    overflowY: "auto",
-    overflowX: "hidden",
-  },
-
-  ".sp-wrapper": {
-    "--sp-colors-fg-active": "#1f2933",
-    "--sp-colors-fg-default": "#757678",
-    "--sp-colors-fg-inactive": "#e4e7eb",
-    "--sp-colors-bg-active": "#e4e7eb",
-    "--sp-colors-bg-default": "#f8f9fb",
-    "--sp-colors-bg-default-overlay": "rgba(248,249,251,0.8117647058823529)",
-    "--sp-colors-bg-input": "#fff",
-    "--sp-colors-accent": "#64d2ff",
-    "--sp-colors-bg-error": "#ffcdca",
-    "--sp-colors-fg-error": "#811e18",
-    "--sp-layout-height": "300px",
-    "--sp-font-size": "14px",
-    "--sp-font-body":
-      '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
-    "--sp-font-mono":
-      '"Fira Mono","DejaVu Sans Mono",Menlo,Consolas,"Liberation Mono",Monaco,"Lucida Console",monospace',
-    "--sp-space-1": "4px",
-    "--sp-space-2": "8px",
-    "--sp-space-3": "12px",
-    "--sp-space-4": "16px",
-    "--sp-space-5": "20px",
-    "--sp-space-6": "24px",
-    "--sp-space-7": "28px",
-    "--sp-space-8": "32px",
-    "--sp-border-radius": "4px",
-    background: "$secondary",
-    fontSize: "var(--sp-font-size)",
-    fontFamily: "var(--sp-font-body)",
-    display: "block",
-    boxSizing: "border-box",
-    textRendering: "optimizeLegibility",
-    WebkitTapHighlightColor: "transparent",
-    WebkitFontSmoothing: "subpixel-antialiased",
-  },
-});
-
 const Canvas = Kit.styled("div", {
   color: "$accessible",
   padding: "$100",
@@ -189,7 +57,7 @@ export default function Playroom({
   const [receivedSource, setSource] = React.useState(source);
   const [code, setCode] = React.useState(thisCode);
   const { resolvedTheme } = useTheme();
-  const [sandboxEmbedTheme, setSandboxEmbedTheme] = React.useState(lightTheme);
+  const [sandboxEmbedTheme, setSandboxEmbedTheme] = React.useState("light");
 
   const Preview = () => {
     const { code: thatCode } = useActiveCode();
@@ -346,13 +214,11 @@ export default function Playroom({
     );
   };
 
-  sandboxGlobalcss();
-
   React.useEffect(() => {
     if (resolvedTheme === "dark") {
-      setSandboxEmbedTheme(darkTheme);
+      setSandboxEmbedTheme("dark");
     } else {
-      setSandboxEmbedTheme(lightTheme);
+      setSandboxEmbedTheme("light");
     }
   }, [resolvedTheme]);
 
@@ -366,8 +232,6 @@ export default function Playroom({
               format: "mdx",
             },
           });
-
-          console.log(event);
 
           setSource(mdxSource);
           setCode(event.data);
@@ -400,16 +264,25 @@ export default function Playroom({
           },
         }}
         theme={sandboxEmbedTheme}
+        style={{ display: "flex", height: "100vh", width: "100%" }}
       >
         <Canvas hasEditor={hasEditor}>
           <Preview />
         </Canvas>
         {hasEditor && (
-          <SandpackLayout>
+          <SandpackLayout
+            style={{
+              border: "1px solid var(--wpds-colors-subtle)",
+              borderRadius: 0,
+              height: "100%",
+              flex: "100%",
+            }}
+          >
             <SandpackCodeEditor
               showLineNumbers
               style={{
                 border: "1px solid var(--wpds-colors-subtle)",
+                height: "100%",
               }}
             />
           </SandpackLayout>