+++ ui/avatar/src/play.stories.jsx
@@ -1,6 +1,5 @@
 import * as React from "react";
 import { Avatar as Component } from "./";
-import Image from "next/image";
 import { styled, theme } from "@washingtonpost/wpds-theme";
 
 export default {
@@ -23,6 +22,8 @@ const Container = styled("div", {
   textAlign: "center",
 });
 const Label = styled("h3", {
+  color: theme.colors.primary,
+  margin: 0,
   textAlign: "center",
 });
 
@@ -33,18 +34,7 @@ export const Avatar = (args) => (
     </Label>
     <Component {...args}>
       <img
-        src="https://i.pravatar.cc/300"
-        alt="An avatar is an atomic component that represents an individual’s identity through a circular photo."
-      />
-    </Component>
-    <Label>Avatar with Element using next/image image</Label>
-    <Component {...args}>
-      <Image
-        loader={myLoader}
-        src="https://i.pravatar.cc/300"
-        width="32"
-        height="32"
-        layout="fixed"
+        src="https://i.pravatar.cc/300?u=test-user@wapo.com"
         alt="An avatar is an atomic component that represents an individual’s identity through a circular photo."
       />
     </Component>