+++ ui/radio-group/src/RadioButton.tsx
@@ -18,10 +18,13 @@ const StyledRadioButton = Theme.styled(RadioGroupPrimitive.Item, {
   borderRadius: "50%",
   borderWidth: "1px",
   cursor: "pointer",
+  padding: 0,
+  transition: Theme.theme.transitions.allFast,
   width: Theme.theme.sizes["125"],
   minWidth: Theme.theme.sizes["125"],
   height: Theme.theme.sizes["125"],
   "&:focus": { borderColor: Theme.theme.colors.cta },
+  "&:focus-visible": { outline: "none" },
   "&:disabled": {
     backgroundColor: Theme.theme.colors.disabled,
     borderColor: Theme.theme.colors.onDisabled,