+++ ui/fieldset/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.14.0](https://github.com/WPMedia/wpds-ui-kit/compare/v0.13.0...v0.14.0) (2022-08-12)
+
+**Note:** Version bump only for package @washingtonpost/wpds-fieldset
+
+
+
+
+
 # [0.13.0](https://github.com/WPMedia/wpds-ui-kit/compare/v0.12.0...v0.13.0) (2022-08-05)
 
 **Note:** Version bump only for package @washingtonpost/wpds-fieldset